feat(schema-type): move types into their own files, add primitive map
