+++ packages/schema-type/package.json
@@ -3,8 +3,8 @@
   "license": "MIT",
   "version": "1.41.0",
   "description": "SchemaType enum for various languages",
-  "main": "src/typescript/SchemaType.ts",
-  "source": "src/typescript/SchemaType.ts",
+  "main": "src/typescript/index.ts",
+  "source": "src/typescript/index.ts",
   "type": "module",
   "repository": {
     "type": "git",
