+++ packages/schema-type/src/typescript/ArraySchemaType.ts
@@ -1,0 +1,101 @@
+import { SchemaType } from "./SchemaType.js";
+
+export type ArraySchemaType =
+  | SchemaType.UINT8_ARRAY
+  | SchemaType.UINT16_ARRAY
+  | SchemaType.UINT24_ARRAY
+  | SchemaType.UINT32_ARRAY
+  | SchemaType.UINT40_ARRAY
+  | SchemaType.UINT48_ARRAY
+  | SchemaType.UINT56_ARRAY
+  | SchemaType.UINT64_ARRAY
+  | SchemaType.UINT72_ARRAY
+  | SchemaType.UINT80_ARRAY
+  | SchemaType.UINT88_ARRAY
+  | SchemaType.UINT96_ARRAY
+  | SchemaType.UINT104_ARRAY
+  | SchemaType.UINT112_ARRAY
+  | SchemaType.UINT120_ARRAY
+  | SchemaType.UINT128_ARRAY
+  | SchemaType.UINT136_ARRAY
+  | SchemaType.UINT144_ARRAY
+  | SchemaType.UINT152_ARRAY
+  | SchemaType.UINT160_ARRAY
+  | SchemaType.UINT168_ARRAY
+  | SchemaType.UINT176_ARRAY
+  | SchemaType.UINT184_ARRAY
+  | SchemaType.UINT192_ARRAY
+  | SchemaType.UINT200_ARRAY
+  | SchemaType.UINT208_ARRAY
+  | SchemaType.UINT216_ARRAY
+  | SchemaType.UINT224_ARRAY
+  | SchemaType.UINT232_ARRAY
+  | SchemaType.UINT240_ARRAY
+  | SchemaType.UINT248_ARRAY
+  | SchemaType.UINT256_ARRAY
+  | SchemaType.INT8_ARRAY
+  | SchemaType.INT16_ARRAY
+  | SchemaType.INT24_ARRAY
+  | SchemaType.INT32_ARRAY
+  | SchemaType.INT40_ARRAY
+  | SchemaType.INT48_ARRAY
+  | SchemaType.INT56_ARRAY
+  | SchemaType.INT64_ARRAY
+  | SchemaType.INT72_ARRAY
+  | SchemaType.INT80_ARRAY
+  | SchemaType.INT88_ARRAY
+  | SchemaType.INT96_ARRAY
+  | SchemaType.INT104_ARRAY
+  | SchemaType.INT112_ARRAY
+  | SchemaType.INT120_ARRAY
+  | SchemaType.INT128_ARRAY
+  | SchemaType.INT136_ARRAY
+  | SchemaType.INT144_ARRAY
+  | SchemaType.INT152_ARRAY
+  | SchemaType.INT160_ARRAY
+  | SchemaType.INT168_ARRAY
+  | SchemaType.INT176_ARRAY
+  | SchemaType.INT184_ARRAY
+  | SchemaType.INT192_ARRAY
+  | SchemaType.INT200_ARRAY
+  | SchemaType.INT208_ARRAY
+  | SchemaType.INT216_ARRAY
+  | SchemaType.INT224_ARRAY
+  | SchemaType.INT232_ARRAY
+  | SchemaType.INT240_ARRAY
+  | SchemaType.INT248_ARRAY
+  | SchemaType.INT256_ARRAY
+  | SchemaType.BYTES1_ARRAY
+  | SchemaType.BYTES2_ARRAY
+  | SchemaType.BYTES3_ARRAY
+  | SchemaType.BYTES4_ARRAY
+  | SchemaType.BYTES5_ARRAY
+  | SchemaType.BYTES6_ARRAY
+  | SchemaType.BYTES7_ARRAY
+  | SchemaType.BYTES8_ARRAY
+  | SchemaType.BYTES9_ARRAY
+  | SchemaType.BYTES10_ARRAY
+  | SchemaType.BYTES11_ARRAY
+  | SchemaType.BYTES12_ARRAY
+  | SchemaType.BYTES13_ARRAY
+  | SchemaType.BYTES14_ARRAY
+  | SchemaType.BYTES15_ARRAY
+  | SchemaType.BYTES16_ARRAY
+  | SchemaType.BYTES17_ARRAY
+  | SchemaType.BYTES18_ARRAY
+  | SchemaType.BYTES19_ARRAY
+  | SchemaType.BYTES20_ARRAY
+  | SchemaType.BYTES21_ARRAY
+  | SchemaType.BYTES22_ARRAY
+  | SchemaType.BYTES23_ARRAY
+  | SchemaType.BYTES24_ARRAY
+  | SchemaType.BYTES25_ARRAY
+  | SchemaType.BYTES26_ARRAY
+  | SchemaType.BYTES27_ARRAY
+  | SchemaType.BYTES28_ARRAY
+  | SchemaType.BYTES29_ARRAY
+  | SchemaType.BYTES30_ARRAY
+  | SchemaType.BYTES31_ARRAY
+  | SchemaType.BYTES32_ARRAY
+  | SchemaType.BOOL_ARRAY
+  | SchemaType.ADDRESS_ARRAY;
