+++ packages/schema-type/src/typescript/DynamicSchemaType.ts
@@ -1,0 +1,4 @@
+import { SchemaType } from "./SchemaType.js";
+import { ArraySchemaType } from "./ArraySchemaType.js";
+
+export type DynamicSchemaType = ArraySchemaType | SchemaType.BYTES | SchemaType.STRING;
