+++ packages/schema-type/src/typescript/StaticSchemaType.ts
@@ -1,0 +1,4 @@
+import { SchemaType } from "./SchemaType.js";
+import { DynamicSchemaType } from "./DynamicSchemaType.js";
+
+export type StaticSchemaType = Exclude<SchemaType, DynamicSchemaType>;
