+++ packages/schema-type/src/typescript/index.ts
@@ -1,0 +1,10 @@
+export type { StaticSchemaType } from "./StaticSchemaType.js";
+export type { DynamicSchemaType } from "./DynamicSchemaType.js";
+export type { ArraySchemaType } from "./ArraySchemaType.js";
+export type { SchemaTypeToPrimitive } from "./SchemaTypeToPrimitive";
+
+export { SchemaType } from "./SchemaType.js";
+export { encodeSchema } from "./encodeSchema.js";
+export { getStaticByteLength } from "./getStaticByteLength.js";
+export { SchemaTypeArrayToElement } from "./SchemaTypeArrayToElement.js";
+export { SchemaTypeId } from "./SchemaTypeId.js";
